+++ src/App.jsx
@@ -4,10 +4,11 @@ import { UsersProvider } from "./component/contexts/usersContext.jsx";
 import MainPage from "./component/page/MainPage.jsx";
 import LogIn from "./component/page/users/LogIn.jsx";
 import SignUp from "./component/page/users/SignUp.jsx";
-import Class from "./component/page/Class.jsx";
+import ClassList from "./component/page/ClassList.jsx";
 import Create from "./component/page/class/Create.jsx";
 import Setting from "./component/page/class/Setting.jsx";
 import Participate from "./component/page/Participate.jsx";
+import Class from "./component/page/Class.jsx";
 import ClassOverview from "./component/page/class/Overview.jsx";
 import ClassSummary from "./component/page/class/Summary.jsx";
 import ClassStudents from "./component/page/class/Students.jsx";
@@ -34,7 +35,7 @@ function App() {
             <Route path="/signup" element={<SignUp />} />
 
             {/* 강의실 페이지 */}
-            <Route path="/class" element={<Class />} />
+            <Route path="/class/list" element={<ClassList />} />
             <Route path="/class/create" element={<Create />} />
             <Route path="/admin/setting" element={<Setting />} />
 
@@ -42,44 +43,19 @@ function App() {
             <Route path="/class/participate" element={<Participate />} />
 
             {/* 강의실 상세 페이지 */}
-            <Route
-              path="/class/:courseId/overview/info"
-              element={<ClassOverview />}
-            />
-            <Route
-              path="/class/:courseId/overview/info/edit"
-              element={<EditClass />}
-            />
-            <Route
-              path="/class/:courseId/overview/notice"
-              element={<ClassNotice />}
-            />
-            {/* 공지사항 생성 페이지 */}
-            <Route
-              path="/class/:courseId/overview/notice/create"
-              element={<NoticeCreate />}
-            />
-            {/* 공지사항 수정 페이지 */}
-            <Route
-              path="/class/:courseId/overview/notice/edit/:noticeId"
-              element={<NoticeCreate />}
-            />
-            <Route
-              path="/class/:courseId/curriculum"
-              element={<ClassCurriculum />}
-            />
-            <Route
-              path="/class/:courseId/curriculum/edit"
-              element={<ClassCurriculumEdit />}
-            />
-            <Route
-              path="/class/:courseId/admin/summary"
-              element={<ClassSummary />}
-            />
-            <Route
-              path="/class/:courseId/admin/students"
-              element={<ClassStudents />}
-            />
+
+            {/* ✅ 강의실 관련 라우트 */}
+          <Route path="/class/:courseId" element={<Class />}>
+            <Route path="overview/info" element={<ClassOverview />} />
+            <Route path="overview/info/edit" element={<EditClass />} />
+            <Route path="overview/notice" element={<ClassNotice />} />
+            <Route path="overview/notice/create" element={<NoticeCreate />} />
+            <Route path="overview/notice/edit/:noticeId" element={<NoticeCreate />} />
+            <Route path="curriculum" element={<ClassCurriculum />} />
+            <Route path="curriculum/edit" element={<ClassCurriculumEdit />} />
+            <Route path="admin/summary" element={<ClassSummary />} />
+            <Route path="admin/students" element={<ClassStudents />} />
+          </Route>
 
             {/* 대시보드 페이지 */}
             <Route path="/dashboard" element={<Dashboard />} />
